+++ R/holiday.R
@@ -22,6 +22,12 @@ globalVariables("holiday")
 #' holiday_aus(2016, state = "VIC")
 #' holiday_aus(2013:2016, state = "ACT")
 holiday_aus <- function(year, state = "national") {
+  if (!requireNamespace("timeDate", quietly = TRUE)) {
+    abort(
+      "Package timeDate required for holiday_aus()", ".\n",
+      "Please install and try again."
+    )
+  }
   if (!is.numeric(year)) {
     abort("`year` must be double/integer.")
   }